+++ edge.py
@@ -15,7 +15,7 @@ from utils.quantization import quantization_nne
 
 
 def cast_to_range(values, scale):
-    return torch.round(values * scale).to(torch.int32) 
+    return torch.round(values * scale).to(torch.long) 
 
 def uncast_from_range(scaled_values, scale):
     return scaled_values / scale
